Test if subtracting autofluorescence from reference spectra would yeild better results
In this version is cubtract the autofluorescence spectra from acceptor and donor reference spectra. But obtain negative values (this is because we ose empirical bacground instead of camera offset). I replace those negative values with zeros. this gives bad results.
Also I've tested with negatives and had good results (some smooth curves). But did not commit that, maybe next time (: